Update README.md
Sentiment analysis (SA) or opinion mining (OM) is a Natural Language Processing (NLP) technique to determine the sentiment or opinion of a given text. A SA model can predict whether a given text data is positive, negative, or neutral by extracting meaning from the natural language.
Social media is an indispensable necessity for modern life. As a result, it is full of people's opinions, emotions, ideas, and attitudes, whether positive or negative. This abundance of views creates many opportunities to apply SA to the education sector, reflecting how countries and cultures develop. in this project , I implemented different machine learning algorithms on a real-world Twitter dataset that contains approximately 8234 Arabic tweets related to Qassim University.
The following this has been done:

1- applying machine learning algorithms (Support vector machines, Naïve Bayes, and k nearest neighbors).
2- Used Python libraries such as NLTK and PyArabic.
3- Describing the dataset from the statistical perspective.
4- Employing different performance measurements(ex. F1-score ,Recall, Accurcy,precision ).
5- showed results of different machine learning algorithms (by confusion matrix ).
6- improve the results. (i used smote).
Therefore, this project's main aim is to use supervised machine learning classifiers to determine the writer's opinion in the tweet (positive, negative, or neutral).